+++ R/fitted.R
@@ -17,6 +17,9 @@ fitted_internal.brmsdraws <- function(draws, scale = "response",
   dpars <- names(draws$dpars)
   if (!length(dpar)) {
     if (scale == "response") {
+      for (nlp in names(draws$nlpars)) {
+        draws$nlpars[[nlp]] <- get_nlpar(draws, nlpar = nlp)
+      }
       for (dp in dpars) {
         draws$dpars[[dp]] <- get_dpar(draws, dpar = dp)
       }