+++ R/priors.R
@@ -563,6 +563,15 @@ prior_effects.brmsterms <- function(x, data, ...) {
     }
     prior <- prior + dp_prior
   }
+  for (nlp in names(x$nlpars)) {
+    nlp_prior <- prior_effects(
+      x$nlpars[[nlp]], data = data,
+      def_scale_prior = def_scale_prior,
+      def_dprior = def_dprior, 
+      spec_intercept = FALSE
+    )
+    prior <- prior + nlp_prior
+  }
   # global population-level priors for categorical models
   if (is_categorical(x$family)) {
     for (cl in c("b", "Intercept")) {
@@ -612,19 +621,7 @@ prior_effects.btl <- function(x, data, spec_intercept = TRUE,
 
 #' @export
 prior_effects.btnl <- function(x, data, ...) {
-  # collect default priors for non-linear parameters
-  # Args:
-  #   see prior_effects.btl
-  nlpars <- names(x$nlpars)
-  prior <- empty_brmsprior()
-  for (i in seq_along(nlpars)) {
-    prior_eff <- prior_effects(
-      x$nlpars[[i]], data = data, 
-      spec_intercept = FALSE, ...
-    )
-    prior <- prior + prior_eff
-  }
-  prior
+  empty_brmsprior()
 }
 
 prior_fe <- function(bterms, data, spec_intercept = TRUE, def_dprior = "") {
@@ -1216,6 +1213,11 @@ check_prior_special.brmsterms <- function(x, prior = NULL, ...) {
       x$dpars[[dp]], prior, allow_autoscale = allow_autoscale, ...
     )
   }
+  for (nlp in names(x$nlpars)) {
+    prior <- check_prior_special(
+      x$nlpars[[nlp]], prior, is_nlpar = TRUE, ...
+    )
+  }
   # copy over the global population-level prior in categorical models
   if (is_categorical(x$family)) {
     for (cl in c("b", "Intercept")) {
@@ -1248,34 +1250,28 @@ check_prior_special.brmsterms <- function(x, prior = NULL, ...) {
 
 #' @export
 check_prior_special.btnl <- function(x, prior, ...) {
-  stopifnot(is.brmsprior(prior))
-  for (nlp in names(x$nlpars)) {
-    prior <- check_prior_special(
-      x$nlpars[[nlp]], prior, is_nlpar = TRUE, ...
-    )
-  }
   prior
 }
 
 #' @export
-check_prior_special.btl <- function(x, prior, data, is_nlpar = FALSE, 
+check_prior_special.btl <- function(x, prior, data,
                                     check_nlpar_prior = TRUE,
                                     allow_autoscale = TRUE, ...) {
   # prepare special priors that cannot be passed to Stan as is
   # Args:
   #   prior: an object of class brmsprior
   #   allow_autoscale: allow autoscaling using sigma?
-  #   is_nlpar: is the parameter to be checked nonlinear?
+  #   check_nlpar_prior: check for priors on non-linear parameters?
   # Returns:
   #   a possibly amended brmsprior object with additional attributes
   px <- check_prefix(x)
-  if (is_nlpar && check_nlpar_prior) {
+  if (is_nlpar(x) && check_nlpar_prior) {
     nlp_prior <- subset2(prior, ls = px)
     if (!any(nzchar(nlp_prior$prior))) {
       stop2(
         "Priors on population-level effects are required in ",
         "non-linear models, but none were found for parameter ", 
-        "'", px$nlp, "'. See help(set_prior) for more details."
+        "'", px$nlpar, "'. See help(set_prior) for more details."
       )
     }
   }