+++ R/validate.R
@@ -79,13 +79,12 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
   # copy stuff from the formula to parameter 'mu'
   str_rhs_form <- formula2str(rhs(formula))
   rhs_needed <- FALSE
-  att <- c("nl", "loop")
   if (is.mixfamily(family)) {
     for (i in seq_along(family$mix)) {
       mui <- paste0("mu", i)
       if (!is.formula(x$pforms[[mui]])) {
         x$pforms[[mui]] <- eval2(paste0(mui, str_rhs_form))
-        attributes(x$pforms[[mui]])[att] <- attributes(formula)[att]
+        attributes(x$pforms[[mui]]) <- attributes(formula)
         rhs_needed <- TRUE
       }
     }
@@ -93,14 +92,14 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
     for (dp in x$family$dpars) {
       if (!is.formula(x$pforms[[dp]])) {
         x$pforms[[dp]] <- eval2(paste0(dp, str_rhs_form))
-        attributes(x$pforms[[dp]])[att] <- attributes(formula)[att]
+        attributes(x$pforms[[dp]]) <- attributes(formula)
         rhs_needed <- TRUE
       }
     }
   } else {
     if (!is.formula(x$pforms[["mu"]])) { 
       x$pforms$mu <- eval2(paste0("mu", str_rhs_form))
-      attributes(x$pforms$mu)[att] <- attributes(formula)[att]
+      attributes(x$pforms$mu) <- attributes(formula)
       rhs_needed <- TRUE
     }
     x$pforms <- move2start(x$pforms, "mu")
@@ -120,22 +119,14 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
   dpars <- names(x$pforms)[dpars]
   dpar_forms <- x$pforms[dpars]
   nlpars <- setdiff(names(x$pforms), dpars)
-  nlpar_forms <- x$pforms[nlpars]
-  dpar_of_nlpars <- rep(NA, length(nlpars))
-  for (i in seq_along(nlpar_forms)) {
-    dpar <- attr(nlpar_forms[[i]], "dpar", TRUE)
-    if (length(dpar) != 1L) {
-      stop2("Parameter '", nlpars[i], "' is not part of the model")
-    }
-    dpar_of_nlpars[i] <- dpar
-  }
+  
+  y$dpars <- named_list(dpars)
   for (dp in dpars) {
     if (get_nl(dpar_forms[[dp]])) {
-      if (is.mixfamily(family) || is_ordinal(family)) {
-        stop2("Non-linear formulas are not yet allowed for this family.")
+      if (is_ordinal(family)) {
+        stop2("Non-linear formulas are not yet allowed in ordinal models.")
       }
-      dpar_nlpar_forms <- nlpar_forms[dpar_of_nlpars %in% dp]
-      y$dpars[[dp]] <- parse_nlf(dpar_forms[[dp]], dpar_nlpar_forms, dp, resp)
+      y$dpars[[dp]] <- parse_nlf(dpar_forms[[dp]], nlpars, resp)
     } else {
       y$dpars[[dp]] <- parse_lf(dpar_forms[[dp]], family = family)
     }
@@ -143,6 +134,34 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
     y$dpars[[dp]]$dpar <- dp
     y$dpars[[dp]]$resp <- resp
   }
+  
+  y$nlpars <- named_list(nlpars)
+  if (length(nlpars)) {
+    nlpar_forms <- x$pforms[nlpars]
+    for (nlp in nlpars) {
+      if (get_nl(nlpar_forms[[nlp]])) {
+        y$nlpars[[nlp]] <- parse_nlf(nlpar_forms[[nlp]], nlpars, resp)
+      } else {
+        y$nlpars[[nlp]] <- parse_lf(nlpar_forms[[nlp]], family = family)
+      }
+      y$nlpars[[nlp]]$nlpar <- nlp
+      y$nlpars[[nlp]]$resp <- resp
+    }
+    used_nlpars <- ulapply(c(y$dpars, y$nlpars), "[[", "used_nlpars")
+    unused_nlpars <- setdiff(nlpars, used_nlpars)
+    if (length(unused_nlpars)) {
+      stop2(
+        "The parameter '", unused_nlpars[1], "' is not a ", 
+        "valid distributional or non-linear parameter. ",
+        "Did you forget to set 'nl = TRUE'?"
+      )
+    }
+    # sort non-linear parameters after dependency
+    used_nlpars <- lapply(y$nlpars, "[[", "used_nlpars")
+    sorted_nlpars <- sort_dependencies(used_nlpars)
+    y$nlpars <- y$nlpars[sorted_nlpars]
+  }
+  
   # fixed distributional parameters
   valid_dpars <- valid_dpars(y)
   inv_fixed_dpars <- setdiff(names(x$pfix), valid_dpars)
@@ -190,6 +209,7 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
   lformula <- c(
     lhsvars, advars, 
     lapply(y$dpars, "[[", "allvars"), 
+    lapply(y$nlpars, "[[", "allvars"), 
     y$time$allvars
   )
   y$allvars <- allvars_formula(lformula)
@@ -251,56 +271,25 @@ parse_lf <- function(formula, family = NULL) {
   structure(y, class = "btl")
 }
 
-parse_nlf <- function(formula, nlpar_forms, dpar = "mu", 
-                      resp = "", is_nlpar = FALSE) {
+parse_nlf <- function(formula, nlpars, resp = "") {
   # parse non-linear formulas
   # Args:
   #   formula: formula of the non-linear model
-  #   nlpar_forms: a list for formulas specifying linear 
-  #     predictors for non-linear parameters
-  #   is_nlpar: Is this a nested non-linear formula?
+  #   nlpars: names of all non-linear parameters
   # Returns:
   #   object of class 'btnl'
-  stopifnot(is.list(nlpar_forms))
-  stopifnot(length(dpar) == 1L)
-  if (!length(nlpar_forms)) {
+  if (!length(nlpars)) {
     stop2("No non-linear parameters specified.")
   }
-  is_nlpar <- as_one_logical(is_nlpar)
   loop <- !isFALSE(attr(formula, "loop"))
   formula <- rhs(as.formula(formula))
   y <- nlist(formula)
-  nlpars <- names(nlpar_forms)
-  y$nlpars <- named_list(nlpars)
-  if (!is_nlpar) {
-    for (nlp in nlpars) {
-      if (get_nl(nlpar_forms[[nlp]])) {
-        take <- !names(nlpar_forms) %in% nlp
-        y$nlpars[[nlp]] <- parse_nlf(
-          nlpar_forms[[nlp]], nlpar_forms[take], is_nlpar = TRUE
-        )     
-      } else {
-        y$nlpars[[nlp]] <- parse_lf(nlpar_forms[[nlp]]) 
-      }
-      y$nlpars[[nlp]]$nlpar <- nlp
-      y$nlpars[[nlp]]$dpar <- dpar
-      y$nlpars[[nlp]]$resp <- resp
-    }
-  }
-  covars <- setdiff(all.vars(formula), nlpars)
-  covars <- unique(c(covars, ulapply(y$nlpars, "[[", "covars")))
+  all_vars <- all.vars(formula)
+  y$used_nlpars <- intersect(all_vars, nlpars)
+  covars <- setdiff(all_vars, nlpars)
   y$covars <- structure(str2formula(covars), rsv_intercept = TRUE)
-  lformula <- c(covars, lapply(y$nlpars, "[[", "allvars"))
-  y$allvars <- allvars_formula(lformula)
+  y$allvars <- allvars_formula(covars)
   environment(y$allvars) <- environment(formula)
-  for (nlp in names(y$nlpars)) {
-    # make sure that nested non-linear parameters 
-    # are aware of all data variables
-    if (is.btnl(y$nlpars[[nlp]])) {
-      y$nlpars[[nlp]]$covars <- y$covars
-      y$nlpars[[nlp]]$allvars <- y$allvars
-    }
-  }
   y$loop <- loop
   structure(y, class = "btnl")
 }
@@ -661,11 +650,12 @@ check_prefix <- function(x, keep_mu = FALSE) {
   x
 }
 
-combine_prefix <- function(prefix, keep_mu = FALSE) {
+combine_prefix <- function(prefix, keep_mu = FALSE, nlp = FALSE) {
   prefix <- check_prefix(prefix, keep_mu = keep_mu)
-  prefix <- lapply(prefix, function(x)
-    ifelse(nzchar(x), paste0("_", x), "")
-  )
+  if (is_nlpar(prefix) && nlp) {
+    prefix$dpar <- "nlp"
+  }
+  prefix <- lapply(prefix, usc)
   sub("^_", "", do.call(paste0, prefix))
 }
 
@@ -716,7 +706,7 @@ plus_rhs <- function(x) {
 }
 
 is_nlpar <- function(x) {
-  !is.null(x[["nlpar"]])
+  isTRUE(nzchar(x[["nlpar"]]))
 }
 
 get_effect <- function(x, ...) {
@@ -741,12 +731,14 @@ get_effect.brmsterms <- function(x, target = "fe", all = TRUE, ...) {
   #   target: type of effects to return
   #   all: logical; include effects of nlpars and dpars?
   if (all) {
-    out <- named_list(names(x$dpars))
-    for (dp in names(out)) {
+    out <- named_list(c(names(x$dpars), names(x$nlpars)))
+    for (dp in names(x$dpars)) {
       out[[dp]] <- get_effect(x$dpars[[dp]], target = target)
     }
+    for (nlp in names(x$nlpars)) {
+      out[[nlp]] <- get_effect(x$nlpars[[nlp]], target = target)
+    }
   } else {
-    x$dpars$mu$nlpars <- NULL
     out <- get_effect(x$dpars[["mu"]], target = target)
   }
   unlist(out, recursive = FALSE)
@@ -759,11 +751,7 @@ get_effect.btl <- function(x, target = "fe", ...) {
 
 #' @export
 get_effect.btnl <- function(x, target = "fe", ...) {
-  out <- named_list(names(x$nlpars))
-  for (nlp in names(out)) {
-    out[[nlp]] <- get_effect(x$nlpars[[nlp]], target = target)
-  }
-  rmNULL(out)
+  NULL
 }
 
 get_advars <- function(x, ...) {
@@ -950,7 +938,7 @@ validate_terms <- function(x) {
   }
   x
 }
-
+ 
 has_intercept <- function(formula) {
   # checks if the formula contains an intercept
   # can handle non-linear formulas