+++ tests/testthat/tests.brmsformula.R
@@ -12,8 +12,6 @@ test_that("brmsformula validates formulas of non-linear parameters", {
 test_that("brmsformula validates formulas of auxiliary parameters", {
   expect_error(bf(y ~ a, ~ 1, sigma ~ 1),
                "Additional formulas must be named")
-  expect_error(bf(y ~ a^x, a ~ 1, family = gaussian()),
-               "The parameter 'a' is not a valid distributional")
 })
 
 test_that("brmsformula does not change a 'brmsformula' object", {